Create custom class-validator validator for Stellar Public Keys
